Stop serializing duplicate fields from the delecate ThrowableProblem
We have a delegate ThrowableProblem which we use @JsonUnwrapped to pull into our response

We then delegate getters for known fields (I assume to maintain consistency should the
ThrowableProblem fields ever change name.

However we don't exclude those delegated fields from the @JsonUnwrapped field so we end
up with them twice in the response.  This wasn't detected before as the duplicates are
silently dropped when converting back to a Map in the tests.

Fixed this by ignoring the fields we delegate, and added a test to check for duplicate
keys in the response string.

Fixes #120